strava_api/requirements.txt
requests
dash
dash-bootstrap-components
pandas
numpy
gunicorn
flask
sqlalchemy
flask-sqlalchemy
flask-login
Werkzeug
configparser
pyyaml
loguru
openpyxl
Flask-WTF
Flask-Session
flask-migrate
Flask-Script
flask-moment
email-validator
psycopg2-binary
flask-mail
pyjwt
python-dotenv
flask-admin
flask-security
libsass
wtforms-sqlalchemy